+++ src/main/java/com/jm/jobseekerplatform/JobseekerPlatformApplication.java
@@ -13,7 +13,7 @@ public class JobseekerPlatformApplication {
     }
 
     //delete "//" to enable init userroles and users to base
-    // @Bean(initMethod = "initData")
+    //@Bean(initMethod = "initData")
     public InitData initialData() {
         return new InitData();
     }
