+++ src/main/java/com/jm/jobseekerplatform/dao/AbstractDAO.java
@@ -8,9 +8,9 @@ import java.lang.reflect.ParameterizedType;
 import java.util.List;
 
 @Repository("abstractDAO")
-public abstract class AbstractDAO<T extends Serializable> {
+public abstract class AbstractDAO<T extends Serializable>  {
 
-    private Class<T> clazz;
+    protected Class<T> clazz;
 
     @PersistenceContext
     protected EntityManager entityManager;
