+++ src/main/java/com/jm/jobseekerplatform/dao/impl/tokens/PasswordResetTokenDao.java
@@ -1,0 +1,9 @@
+//package com.jm.jobseekerplatform.dao.impl.tokens;
+//
+//import com.jm.jobseekerplatform.model.tokens.PasswordResetToken;
+//import org.springframework.stereotype.Repository;
+//
+//@Repository
+//public class PasswordResetTokenDao extends BaseTokenDAO<PasswordResetToken> {
+//
+//}
