+++ src/main/java/com/jm/jobseekerplatform/dao/impl/tokens/VerificationTokenDAO.java
@@ -1,0 +1,9 @@
+//package com.jm.jobseekerplatform.dao.impl.tokens;
+//
+//        import com.jm.jobseekerplatform.model.tokens.VerificationToken;
+//        import org.springframework.stereotype.Repository;
+//
+//@Repository
+//public class VerificationTokenDAO extends BaseTokenDAO<VerificationToken> {
+//
+//}
